+++ tests/test.py
@@ -6,13 +6,15 @@ from os.path import dirname
 
 if __name__ == '__main__':
     here = dirname(__file__)
-    sys.path.insert(0, here+'/..')
+    sys.path.insert(0, here + '/..')
 
 import freesms
+
 
 class TestVersion(unittest.TestCase):
     def test_version(self):
         self.assertRegexpMatches(freesms.__version__, r'^\d+\.\d+\.\d+')
+
 
 if __name__ == '__main__':
     suite = unittest.defaultTestLoader.discover(here)